Add ResilientSMTPChannel
A spam email was observed that contained an invalid octet, so the
Python smtpd module caused an uncaught UnicodeDecodeError to be raised
when trying to decode the input as UTF-8.

Catch this and log it instead of dropping it on the floor.

As an added benefit, log SMTPChannel messages using the logging module
instead of writing them to the tty (which is the default).