Change handle_error interface to fix fatal bug
When an exception occurred in the Message() constructor,
the try-clause of handle_message would error out before message was
assigned a value. Then the except-clause would try to use message
anyway, resulting in an error.

This means that handle_error would not even be called.
Therefore, we change the interface to handle_error:
The message may be None, and an additional parameter, str_data,
is passed. This way, handle_error can still log errors caused by bugs
in emailtunnel's Message parsing.